Income expense

+++ database/migrations/2024_05_31_120535_create_banking_records_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('bank_name');
             $table->string('account_number')->unique();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->float('balance')->default(0);
+            $table->decimal('balance', 15, 2)->default(0);
             $table->enum('valuta', ['EUR', 'USD'])->default('EUR')->nullable(); //TODO: enum, welke valuta gaan we gebruiken?
             $table->float('exchange_rate')->nullable(); //TODO: Hoe implementeren?
             $table->timestamps();
